more changes

+++ onnxruntime/core/providers/cpu/math/matmul_helper.h
@@ -31,9 +31,9 @@ class MatMulComputeHelper {
     // A: [M1, M2, ... K], B: [1, ..., 1, N, K]^T
     if (!transa && left_num_dims >= 2 && right_num_dims >= 2 && left_num_dims >= right_num_dims &&
         right_shape.SizeToDimension(right_num_dims - 1) == right_shape[right_num_dims - 2]) {
-      M_ = left_shape.SizeToDimension(left_num_dims - 1);
-      K_ = left_shape[left_num_dims - 1];
-      N_ = transb ? right_shape[right_num_dims - 2] : right_shape[right_num_dims - 1];
+      M_ = gsl::narrow<ptrdiff_t>(left_shape.SizeToDimension(left_num_dims - 1));
+      K_ = gsl::narrow<ptrdiff_t>(left_shape[left_num_dims - 1]);
+      N_ = gsl::narrow<ptrdiff_t>(transb ? right_shape[right_num_dims - 2] : right_shape[right_num_dims - 1]);
       output_shape_ = left_shape;
       output_shape_[left_num_dims - 1] = N_;
       output_offsets_ = {0};
@@ -55,18 +55,18 @@ class MatMulComputeHelper {
     // output shape would squeeze the reduced 1D dimension
     size_t num_output_dims = num_input_dims - (has_1D_input ? 1 : 0);
 
-    left_padded_dims_ = std::vector<int64_t>(num_dims_with_pad, 1);
-    right_padded_dims_ = std::vector<int64_t>(num_dims_with_pad, 1);
+    left_padded_dims_ = std::vector<ptrdiff_t>(num_dims_with_pad, 1);
+    right_padded_dims_ = std::vector<ptrdiff_t>(num_dims_with_pad, 1);
 
     if (right_num_dims == 1) {
       // right padded to (1,...,K,1)
-      right_padded_dims_[num_dims_with_pad - 2] = right_shape[0];
+      right_padded_dims_[num_dims_with_pad - 2] = static_cast<ptrdiff_t>(right_shape[0]);
 
       if (num_input_dims >= 2) {
         // left padded to (...,1,K)
         left_shape.CopyDims(&left_padded_dims_[0], left_num_dims - 2);
-        left_padded_dims_[num_dims_with_pad - 3] = left_shape[transa ? left_num_dims - 1 : left_num_dims - 2];
-        left_padded_dims_[num_dims_with_pad - 1] = left_shape[transa ? left_num_dims - 2 : left_num_dims - 1];
+        left_padded_dims_[num_dims_with_pad - 3] = static_cast<ptrdiff_t>(left_shape[transa ? left_num_dims - 1 : left_num_dims - 2]);
+        left_padded_dims_[num_dims_with_pad - 1] = static_cast<ptrdiff_t>(left_shape[transa ? left_num_dims - 2 : left_num_dims - 1]);
       } else {
         // pad 1 in the front
         left_shape.CopyDims(&left_padded_dims_[num_dims_with_pad - left_num_dims], left_num_dims);
@@ -90,17 +90,17 @@ class MatMulComputeHelper {
         ORT_RETURN_IF_NOT(right_padded_dims_[idx_dim] == 1, "right operand cannot broadcast on dim ", idx_dim);
     }
     if (transa) {
-      M_ = has_1D_input ? 1 : left_shape[left_num_dims - 1];
-      K_ = left_shape[left_num_dims - 2];
+      M_ = gsl::narrow<ptrdiff_t>(has_1D_input ? 1 : left_shape[left_num_dims - 1]);
+      K_ = gsl::narrow<ptrdiff_t>(left_shape[left_num_dims - 2]);
     } else {
-      M_ = has_1D_input ? 1 : left_shape[left_num_dims - 2];
-      K_ = left_shape[left_num_dims - 1];
+      M_ = gsl::narrow<ptrdiff_t>(has_1D_input ? 1 : left_shape[left_num_dims - 2]);
+      K_ = gsl::narrow<ptrdiff_t>(left_shape[left_num_dims - 1]);
     }
 
     if (transb) {
-      N_ = (right_num_dims == 1) ? 1 : right_shape[right_num_dims - 2];
+      N_ = gsl::narrow<ptrdiff_t>((right_num_dims == 1) ? 1 : right_shape[right_num_dims - 2]);
     } else {
-      N_ = (right_num_dims == 1) ? 1 : right_shape[right_num_dims - 1];
+      N_ = gsl::narrow<ptrdiff_t>((right_num_dims == 1) ? 1 : right_shape[right_num_dims - 1]);
     }
 
     if (!has_1D_input) {
@@ -181,7 +181,7 @@ class MatMulComputeHelper {
       auto left_dim = left_padded_dims_[idx_dim];
       auto right_dim = right_padded_dims_[idx_dim];
       auto output_dim = output_broadcast_dims_[idx_dim];
-      for (int64_t i = 0; i < output_dim; ++i) {
+      for (int i = 0; i < output_dim; ++i) {
         RecursiveFill(idx_dim + 1,
                       idx_left + i * (left_dim == 1 ? 0 : left_padded_strides_[idx_dim]),
                       idx_right + i * (right_dim == 1 ? 0 : right_padded_strides_[idx_dim]),
@@ -197,9 +197,9 @@ class MatMulComputeHelper {
 
   size_t num_broadcasted_dims_;
 
-  std::vector<int64_t> left_padded_dims_;
-  std::vector<int64_t> right_padded_dims_;
-  std::vector<int64_t> output_broadcast_dims_;
+  std::vector<ptrdiff_t> left_padded_dims_;
+  std::vector<ptrdiff_t> right_padded_dims_;
+  std::vector<ptrdiff_t> output_broadcast_dims_;
 
   std::vector<size_t> left_padded_strides_;
   std::vector<size_t> right_padded_strides_;
@@ -207,9 +207,9 @@ class MatMulComputeHelper {
 
   TensorShape output_shape_;
 
-  int64_t M_;
-  int64_t N_;
-  int64_t K_;
+  ptrdiff_t M_;
+  ptrdiff_t N_;
+  ptrdiff_t K_;
 
   std::vector<size_t> left_offsets_;
   std::vector<size_t> right_offsets_;
@@ -222,17 +222,17 @@ class MatMulComputeHelper {
   }
 
   // left and output matrices' first dim
-  int64_t M() const {
+  ptrdiff_t M() const {
     return M_;
   }
 
   // right and output matrices' second dim
-  int64_t N() const {
+  ptrdiff_t N() const {
     return N_;
   }
 
   // left matrices' second dim, and right matrices' first dim
-  int64_t K() const {
+  ptrdiff_t K() const {
     return K_;
   }
